sim-server/requirements.txt
appdirs==1.4.3
astroid==2.3.3
certifi==2019.11.28
chardet==3.0.4
Click==7.0
colorama==0.4.3
configparser==4.0.2
contextlib2==0.6.0.post1
distlib==0.3.0
enum34==1.1.6
filelock==3.0.12
Flask==1.1.1
idna==2.9
importlib-metadata==1.5.0
importlib-resources==1.0.2
isort==4.3.21
itsdangerous==1.1.0
Jinja2==2.11.1
lazy-object-proxy==1.4.3
MarkupSafe==1.1.1
mccabe==0.6.1
pathlib2==2.3.5
pipenv==2018.11.26
psutil==5.7.0
pylint==2.4.4
requests==2.23.0
scandir==1.10.0
six==1.14.0
typing==3.7.4.1
urllib3==1.25.8
virtualenv==20.0.4
virtualenv-clone==0.5.3
websocket-client==0.57.0
websockets==8.1
Werkzeug==1.0.0
wrapt==1.11.2
zipp==1.2.0
